use-cases/requirements.txt
# Tavily SDK
tavily-python>=0.7.0

# LangChain ecosystem
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Data validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0
httpx>=0.25.0
